requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
langchain-groq==0.1.9
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.1.4
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2
twilio==8.10.0
stripe==7.8.0
redis==5.0.1
celery==5.3.4
prometheus-client==0.19.0
structlog==23.2.0
colorlog==6.8.0
pytest==7.4.3
pytest-asyncio==0.21.1
